+++ lib/config/app_version.dart
@@ -5,5 +5,5 @@
 class AppVersion {
   const AppVersion._();
 
-  static const String text = 'Web App 0.9.57 · r385a';
+  static const String text = 'Web App 0.9.57 · r385b';
 }